refactor: extract DataSource as a component
